Switched to floating point value

+++ Source/Control.cpp
@@ -10,7 +10,7 @@
 
 #include "Control.h"
 
-Control::Control(int id) : val(0),
+Control::Control(int id) : val(0.0),
                            isOn(false),
                            ChangeBroadcaster()
 {
@@ -19,7 +19,7 @@ Control::Control(int id) : val(0),
 
 Control::~Control() = default;
 
-void Control::update(int val, bool isOn)
+void Control::update(float val, bool isOn)
 {
   this->val = val;
   this->isOn = isOn;
